fix: rename package
Renaming the package to be unique will allow Java modules to be used.
