+++ src/main/java/com/alexandriasoftware/jsplitbutton/action/ButtonClickedActionListener.java
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alexandriasoftware.swing.action;
+package com.alexandriasoftware.jsplitbutton.action;
 
-import com.alexandriasoftware.swing.JSplitButton;
+import com.alexandriasoftware.jsplitbutton.JSplitButton;
 import java.awt.event.ActionListener;
 
 /**
