+++ src/main/java/com/alexandriasoftware/jsplitbutton/action/package-info.java
@@ -1,0 +1,20 @@
+/*
+ * Copyright (C) 2016, 2018 Randall Wood
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+/**
+ * {@link java.util.EventListener}s for
+ * {@link com.alexandriasoftware.jsplitbutton.JSplitButton}.
+ */
+package com.alexandriasoftware.jsplitbutton.action;
