+++ src/main/java/com/alexandriasoftware/jsplitbutton/module-info.java
@@ -1,11 +1,11 @@
 /*
- * Copyright (C) 2016, 2018 Randall Wood
+ * Copyright 2022 rhwood.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,8 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/**
- * {@link java.util.EventListener}s for
- * {@link com.alexandriasoftware.swing.JSplitButton}.
- */
-package com.alexandriasoftware.swing.action;
+
+module JSplitButton {
+    exports com.alexandriasoftware.jsplitbutton;
+    requires java.desktop;
+}
